refactor: 커스텀 훅 useData 구현
