+++ src/pages/Section.tsx
@@ -7,26 +7,19 @@ import addImg from 'assets/images/add.png';
 import { borderColor } from 'constants/color';
 import MESSAGE from 'constants/message';
 import PATH from 'constants/PATH';
+import useData from 'hooks/useData';
 import useRedirect from 'hooks/useRedirect';
 import React, { useEffect, useState } from 'react';
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { clearSelectedLIne, getLinesAsync, getSelectedLineAsync } from 'redux/lineSlice';
 import { addSectionAsync, AddSectionPayload, deleteSectionAsync } from 'redux/sectionSlice';
 import { getStationAsync } from 'redux/stationSlice';
-import { RootState } from 'redux/store';
-import { LineInterface, SelectedLineInterface, StationInterface } from 'types';
 import AddSectionModal from './AddSectionModal';
 
 const Section = () => {
   useRedirect(PATH.LOGIN);
-
   const dispatch = useDispatch();
-
-  const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
-  const stations: StationInterface[] | null = useAppSelector((state) => state.station.stations);
-  const lines: LineInterface[] | null = useAppSelector((state) => state.line.lines);
-  const selectedLine: SelectedLineInterface | null = useAppSelector((state) => state.line.selectedLine);
-
+  const { stations, lines, selectedLine } = useData();
   const [selectedLineId, setSelectedLineId] = useState(0);
   const [modalOpen, setModalOpen] = useState(false);
